// frontend/sdk/models/TrackOptionDto.ts
export interface TrackOptionDto {
    /**
     * The ID of the track
     * @type {string}
     * @memberof TrackOptionDto
     */
    id: string;
    /**
     * The name of the track
     * @type {string}
     * @memberof TrackOptionDto
     */
    name: string;
    /**
     * Normalized name for search/matching
     * @type {string}
     * @memberof TrackOptionDto
     */
    normalizedName: string;
    /**
     * The artist of the track
     * @type {string}
     * @memberof TrackOptionDto
     */
    artist: string;
    /**
     * Normalized artist for search/matching
     * @type {string}
     * @memberof TrackOptionDto
     */
    normalizedArtist: string;
    /**
     * Every credited artist, primary first
     * @type {Array<string>}
     * @memberof TrackOptionDto
     */
    allArtists?: Array<string>;
    /**
     * Link to the track in the catalogue it came from
     * @type {string}
     * @memberof TrackOptionDto
     */
    trackUrl?: string;
    /**
     * ISRC of the recording, used to judge the guess
     * @type {string}
     * @memberof TrackOptionDto
     */
    isrc?: string;
    /**
     * The album image URL of the track
     * @type {string}
     * @memberof TrackOptionDto
     */
    albumImageUrl?: string;
    /**
     * The album name of the track
     * @type {string}
     * @memberof TrackOptionDto
     */
    albumName?: string;
    /**
     * The album URL of the track
     * @type {string}
     * @memberof TrackOptionDto
     */
    albumUrl?: string;
    /**
     * The release year of the track
     * @type {number}
     * @memberof TrackOptionDto
     */
    releaseYear?: number;
}